# api/openapi.yaml
openapi: 3.0.3
info:
  title: SnapDeploy Core API
  description: A modern API for the SnapDeploy platform
  version: 1.0.0
  contact:
    name: SnapDeploy Team
    email: support@snapdeploy.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://core-dev.snap-deploy.com/api/v1
    description: Production server
security:
  - CognitoAuth: []
paths:
...
  /auth/me:
    get:
      summary: Get current user information
      description: Returns information about the currently authenticated user
      tags:
        - Authentication
      responses:
        "200":
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /users/{id}/repos/sync:
    post:
      summary: Sync user repositories from GitHub
      description: Syncs the repositories for a user from GitHub
      tags:
        - Repositories
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Repositories synced successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRepositoriesSyncResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{id}/repos:
    get:
      summary: Get user repositories with search
      description: Returns the repositories for a user with pagination and optional search. Search queries match against repository name, full name, and description.
      tags:
        - Repositories
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          description: Page number (default is 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          description: Number of items per page (default is 20, max is 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          required: false
          description: Search query to filter repositories by name, full name, or description
          schema:
            type: string
          example: "react"
      responses:
        "200":
          description: Repositories retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRepositoriesResponse"
              examples:
                with_search:
                  summary: Search results
                  value:
                    repositories:
                      - id: "550e8400-e29b-41d4-a716-446655440000"
                        name: "my-react-app"
                        full_name: "user/my-react-app"
                        description: "A React application"
                        url: "https://api.github.com/repos/user/my-react-app"
                        html_url: "https://github.com/user/my-react-app"
                        private: false
                        fork: false
                        stars: 10
                        watchers: 5
                        forks: 2
                        language: "JavaScript"
                        created_at: "2024-01-15T10:30:00Z"
                    pagination:
                      page: 1
                      limit: 20
                      total: 1
                      total_pages: 1
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{id}/projects:
    get:
      summary: Get user projects
      description: Returns all projects for a user with pagination
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          description: Page number (default is 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          description: Number of items per page (default is 20, max is 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: Create a new project
      description: Creates a new project for a user
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          description: Project with this repository URL already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{id}:
    get:
      summary: Get a project by ID
      description: Returns a single project by its ID
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: Update a project
      description: Updates an existing project
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: You don't have permission to update this project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete a project
      description: Deletes a project
      tags:
        - Projects
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Project deleted successfully
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: You don't have permission to delete this project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /deployments:
    post:
      summary: Create a new deployment
      description: Creates a new deployment for a project
      tags:
        - Deployments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeploymentRequest"
      responses:
        "201":
          description: Deployment created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: You don't have permission to create a deployment for this project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /deployments/{id}:
    get:
      summary: Get a deployment by ID
      description: Returns a single deployment by its ID
      tags:
        - Deployments
      parameters:
        - name: id
          in: path
          required: true
          description: Deployment ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deployment retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete a deployment
      description: Deletes a deployment
      tags:
        - Deployments
      parameters:
        - name: id
          in: path
          required: true
          description: Deployment ID
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deployment deleted successfully
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: You don't have permission to delete this deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /deployments/{id}/status:
    patch:
      summary: Update deployment status
      description: Updates the status of a deployment
      tags:
        - Deployments
      parameters:
        - name: id
          in: path
          required: true
          description: Deployment ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeploymentStatusRequest"
      responses:
        "200":
          description: Deployment status updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: You don't have permission to update this deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /deployments/{id}/logs:
    post:
      summary: Append to deployment logs
      description: Appends a log line to a deployment
      tags:
        - Deployments
      parameters:
        - name: id
          in: path
          required: true
          description: Deployment ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppendDeploymentLogRequest"
      responses:
        "200":
          description: Log appended successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: You don't have permission to update this deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{id}/deployments:
    get:
      summary: Get project deployments
      description: Returns all deployments for a project with pagination
      tags:
        - Deployments
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          description: Page number (default is 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          description: Number of items per page (default is 20, max is 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Deployments retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{id}/deployments/latest:
    get:
      summary: Get latest project deployment
      description: Returns the most recent deployment for a project
      tags:
        - Deployments
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Latest deployment retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: No deployments found for this project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{id}/deployments:
    get:
      summary: Get user deployments
      description: Returns all deployments for a user with pagination
      tags:
        - Deployments
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          description: Page number (default is 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          description: Number of items per page (default is 20, max is 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Deployments retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    CognitoAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT token
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        hasSyncedRepositories:
          type: boolean
          description: Indicates whether the user has synced any repositories
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserRepositoriesSyncResponse:
      type: object
      properties:
        message:
          type: string
    UserRepositoriesResponse:
      type: object
      properties:
        repositories:
          type: array
          items:
            $ref: "#/components/schemas/Repository"
        pagination:
          $ref: "#/components/schemas/Pagination"
    Repository:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Repository unique identifier
        name:
          type: string
          description: Repository name
          example: "my-react-app"
        full_name:
          type: string
          description: Full repository name (owner/repo)
          example: "user/my-react-app"
        description:
          type: string
          nullable: true
          description: Repository description
          example: "A React application for managing tasks"
        url:
          type: string
          description: API URL for the repository
          example: "https://api.github.com/repos/user/my-react-app"
        html_url:
          type: string
          nullable: true
          description: Web URL for the repository
          example: "https://github.com/user/my-react-app"
        private:
          type: boolean
          description: Whether the repository is private
          example: false
        fork:
          type: boolean
          description: Whether the repository is a fork
          example: false
        stars:
          type: integer
          description: Number of stars
          example: 42
        watchers:
          type: integer
          description: Number of watchers
          example: 15
        forks:
          type: integer
          description: Number of forks
          example: 7
        language:
          type: string
          nullable: true
          description: Primary programming language
          example: "JavaScript"
        created_at:
          type: string
          format: date-time
          description: Repository creation timestamp
          example: "2024-01-15T10:30:00Z"
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Number of items per page
        total:
          type: integer
          description: Total number of items
        total_pages:
          type: integer
          description: Total number of pages
    CreateProjectRequest:
      type: object
      required:
        - repository_url
        - install_command
        - build_command
        - run_command
        - language
      properties:
        repository_url:
          type: string
          description: Repository URL
          example: "https://github.com/user/my-app"
        install_command:
          type: string
          description: Install command to execute (installs dependencies)
          example: "npm install"
        build_command:
          type: string
          description: Build command to execute
          example: "npm run build"
        run_command:
          type: string
          description: Run command to execute
          example: "npm start"
        language:
          type: string
          description: Programming language or framework
          enum: [NODE, NODE_TS, NEXTJS, GO, PYTHON]
          example: "NODE_TS"
    UpdateProjectRequest:
      type: object
      required:
        - repository_url
        - install_command
        - build_command
        - run_command
        - language
      properties:
        repository_url:
          type: string
          description: Repository URL
          example: "https://github.com/user/my-app"
        install_command:
          type: string
          description: Install command to execute (installs dependencies)
          example: "npm install"
        build_command:
          type: string
          description: Build command to execute
          example: "npm run build"
        run_command:
          type: string
          description: Run command to execute
          example: "npm start"
        language:
          type: string
          description: Programming language or framework
          enum: [NODE, NODE_TS, NEXTJS, GO, PYTHON]
          example: "NODE_TS"
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Project unique identifier
        user_id:
          type: string
          format: uuid
          description: User ID who owns the project
        repository_url:
          type: string
          description: Repository URL
          example: "https://github.com/user/my-app"
        install_command:
          type: string
          description: Install command to execute (installs dependencies)
          example: "npm install"
        build_command:
          type: string
          description: Build command to execute
          example: "npm run build"
        run_command:
          type: string
          description: Run command to execute
          example: "npm start"
        language:
          type: string
          description: Programming language or framework
          enum: [NODE, NODE_TS, NEXTJS, GO, PYTHON]
          example: "NODE_TS"
        created_at:
          type: string
          format: date-time
          description: Project creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Project last update timestamp
    ProjectListResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        pagination:
          $ref: "#/components/schemas/Pagination"
    CreateDeploymentRequest:
      type: object
      required:
        - project_id
        - commit_hash
        - branch
      properties:
        project_id:
          type: string
          format: uuid
          description: Project ID to deploy
          example: "550e8400-e29b-41d4-a716-446655440000"
        commit_hash:
          type: string
          description: Git commit hash to deploy
          example: "abc123def456"
          minLength: 7
          maxLength: 40
        branch:
          type: string
          description: Git branch name
          example: "main"
          maxLength: 255
    UpdateDeploymentStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          description: New deployment status
          enum: [PENDING, BUILDING, DEPLOYING, DEPLOYED, FAILED, ROLLED_BACK]
          example: "BUILDING"
    AppendDeploymentLogRequest:
      type: object
      required:
        - log_line
      properties:
        log_line:
          type: string
          description: Log line to append
          example: "Building Docker image..."
    Deployment:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Deployment unique identifier
        project_id:
          type: string
          format: uuid
          description: Project ID this deployment belongs to
        user_id:
          type: string
          format: uuid
          description: User ID who created the deployment
        commit_hash:
          type: string
          description: Git commit hash
          example: "abc123def456"
        branch:
          type: string
          description: Git branch name
          example: "main"
        status:
          type: string
          description: Current deployment status
          enum: [PENDING, BUILDING, DEPLOYING, DEPLOYED, FAILED, ROLLED_BACK]
          example: "DEPLOYED"
        logs:
          type: string
          description: Deployment logs
          example: "Starting build process...\nBuilding Docker image...\nDeployment completed successfully!"
        created_at:
          type: string
          format: date-time
          description: Deployment creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Deployment last update timestamp
    DeploymentListResponse:
      type: object
      properties:
        deployments:
          type: array
          items:
            $ref: "#/components/schemas/Deployment"
        pagination:
          $ref: "#/components/schemas/Pagination"
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...